chore: move the useMediatorSetup hook
