+++ packages/agent/src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from './useCredentialsForDisplay'
 export * from './useCredentialForDisplayById'
 export * from './useInboxNotifications'
-export * from './useMediatorSetup'
 export * from './useCredentialByCategory'
