+++ packages/sdk/src/hooks/useMediatorSetup.tsx
@@ -1,13 +1,15 @@
 import { useEffect, useState } from 'react'
-import type { ParadymAppAgent } from '../agent'
-import { hasMediationConfigured, setupMediationWithDid, useMessagePickup } from '../mediation'
+import type { DidCommAgent } from '../agent'
+import { hasMediationConfigured, setupMediationWithDid } from '../didcomm/mediation'
+import { ParadymWalletNoMediatorDidProvidedError } from '../error'
+import { useMessagePickup } from './useMessagePickup'
 
 export function useMediatorSetup({
   agent,
   hasInternetConnection,
   mediatorDid,
 }: {
-  agent?: ParadymAppAgent
+  agent?: DidCommAgent
   hasInternetConnection: boolean
   mediatorDid?: string
 }) {
@@ -17,7 +19,7 @@ export function useMediatorSetup({
   // Enable message pickup when mediation is configured and internet connection is available
   useMessagePickup({
     isEnabled: hasInternetConnection && isMediationConfigured,
-    agent: agent as ParadymAppAgent,
+    agent,
   })
 
   useEffect(() => {
@@ -27,18 +29,17 @@ export function useMediatorSetup({
 
     setIsSettingUpMediation(true)
 
-    agent.config.logger.debug('Checking if mediation is configured.')
+    agent.config.logger.debug('Checking if mediation is configured')
 
     void hasMediationConfigured(agent)
       .then(async (mediationConfigured) => {
         if (!mediationConfigured) {
-          agent.config.logger.debug('Mediation not configured yet.')
-          if (!mediatorDid) throw new Error('No mediator did provided.')
-
+          agent.config.logger.debug('Mediation not configured yet')
+          if (!mediatorDid) throw new ParadymWalletNoMediatorDidProvidedError()
           await setupMediationWithDid(agent, mediatorDid)
         }
 
-        agent.config.logger.info("Mediation configured. You're ready to go!")
+        agent.config.logger.info('Mediation configured')
         setIsMediationConfigured(true)
       })
       .finally(() => {
